API - Refactored request constrution to util
